Revert "Beginning of documentation"
This reverts commit 2cf701687cc768526b66c5ca2efa07e2d44779ec.
