--- src/main/scala/scaldi/injectable/ByWord.scala
@@ -1,8 +1,0 @@
-package scaldi.injectable
-
-import scaldi.Identifier
-
-class ByWord(initialIdentifiers: List[Identifier] = Nil) {
-  def default[T](fn: => T) = new InjectConstraints[T](Some(() => fn), initialIdentifiers)
-}
-
