--- src/main/scala/scaldi/injectable/InjectConstraints.scala
@@ -1,20 +1,0 @@
-package scaldi.injectable
-
-import scaldi.Identifier
-
-case class InjectConstraints[+T](default: Option[() => T] = None, initialIdentifiers: List[Identifier] = Nil) {
-  var identifiers : List[Identifier] = initialIdentifiers
-
-  def and(ids: Identifier*) = {
-    identifiers = identifiers ++ ids
-    this
-  }
-
-  def and(by: ByWord) = new ByWord(identifiers)
-  def which(by: ByWord) = and(by)
-  def that(by: ByWord) = and(by)
-  def is(by: ByWord) = and(by)
-
-  def and(by: IdentifiedWord[_]) = new IdentifiedWord[T](default, identifiers)
-}
-
