+++ src/test/scala/scaldi/InjectorSpec.scala
@@ -1,7 +1,5 @@
 package scaldi
 
-import scaldi.injectable.Injectable
-
 import language.postfixOps
 
 import org.scalatest.{Matchers, WordSpec}
@@ -37,7 +35,7 @@ class InjectorSpec extends WordSpec with Matchers {
     }
 
     "compose injectors so, that injectors that come first should override bindings of later injectors" in {
-      import Injectable._
+      import scaldi.Injectable._
 
       {
         implicit val injector = new Test2Module ++ new Test1Module ++ new AppModule
@@ -121,7 +119,7 @@ class InjectorSpec extends WordSpec with Matchers {
   "SystemPropertiesInjector" should {
     "look for simple bindings in system properties and convert them to required type" in {
       implicit val injector = SystemPropertiesInjector :: new AppModule
-      import Injectable._
+      import scaldi.Injectable._
 
       try {
         System.setProperty("host", "test-sys")
