+++ src/test/scala/scaldi/testData.scala
@@ -1,7 +1,5 @@
 package scaldi
 
-import scaldi.injectable.Injectable
-
 trait Server
 case class HttpServer(host: String, port: Int) extends Server
 class LifecycleServer extends Server {
@@ -31,8 +29,7 @@ class TcpServer(implicit inj: Injector) extends Server with Injectable  {
   def getConnection = new TcpConnection
 }
 
-import scaldi.injectable.Injectable
-import Injectable._
+import scaldi.Injectable._
 
 class TcpConnection(implicit inj: Injector) {
   val welcomeMessage = inject [String] ('welcome is by default "Hi")
